Reduce range per elevation for rainwater
Given that the width (distance between the first and the last point) of
each elevation level cannot increase and is likely to decrease per
elevation, use the start/end boundaries from each preceding level to
reduce the range of points to look at. If the width of each elevation
level does not decrease substantially with elevation this will actually
lead to worse runtime due to additional computations being performed and
larger size of the data referenced by the loop.